Account for extensionless urls in translation workflow

+++ rosey-connector/generateTranslationFiles.js
@@ -1,12 +1,14 @@
 import fs from "fs";
 import YAML from "yaml";
 import path from "path";
-import dotenv, { config } from "dotenv";
+import dotenv from "dotenv";
 import { NodeHtmlMarkdown } from "node-html-markdown";
 import {
   isDirectory,
   readFileWithFallback,
   readJsonFromFile,
+  getTranslationHTMLFilename,
+  getTranslationHTMLFilenameExtensionless,
 } from "./helpers/file-helper.js";
 
 const nhm = new NodeHtmlMarkdown(
@@ -68,14 +70,19 @@ async function generateTranslationFilesForLocale(locale, configData) {
       }
 
       const fileNameHTMLFormatted = getTranslationHTMLFilename(fileNameWithExt);
+      const fileNameHtmlFormattedExtensionlessUrl =
+        getTranslationHTMLFilenameExtensionless(fileNameWithExt);
 
-      if (!pages.includes(fileNameHTMLFormatted)) {
+      if (
+        !pages.includes(fileNameHTMLFormatted) &&
+        !pages.includes(fileNameHtmlFormattedExtensionlessUrl)
+      ) {
         console.log(
-          `❌ Deleting ${fileNameHTMLFormatted}(${filePath}), since it doesn't exist in the pages in our base.json`
+          `🧹 The page ${fileNameWithExt} doesn't exist in the pages in our base.json - deleting!`
         );
 
         await fs.promises.unlink(filePath);
-        console.log(`❌ ${fileNameHTMLFormatted} at ${filePath} was deleted`);
+        console.log(`🧹 Translation file ${filePath} was deleted!`);
       }
     })
   );
@@ -244,8 +251,8 @@ function getInputConfig(inputKey, page, inputTranslationObj, baseURL) {
   const inputType = isKeyMarkdown
     ? "markdown"
     : isInputShortText
-      ? "text"
-      : "textarea";
+    ? "text"
+    : "textarea";
 
   const options = isKeyMarkdown
     ? {
@@ -296,18 +303,6 @@ function getInputConfig(inputKey, page, inputTranslationObj, baseURL) {
       };
 
   return inputConfig;
-}
-
-function getTranslationHTMLFilename(translationFilename) {
-  if (translationFilename === "404.yaml") {
-    return "404.html";
-  }
-
-  if (translationFilename === "home.yaml") {
-    return "index.html";
-  }
-
-  return translationFilename.replace(".yaml", "/index.html");
 }
 
 function generateLocationString(originalPhrase, page, baseURL) {
